Factor out loopback block device API.
